+++ CHANGES.rst
@@ -44,6 +44,7 @@ Added
 - Document how to set up a development environment, run tests, run linters and update
   contributors list in ``CONTRIBUTING.rst``.
 - Document how to pin reformatter/linter versions in ``pre-commit``.
+- Clarify configuration of reformatter/linter tools in README and ``--help``.
 
 Fixed
 -----